Create Simple Regression approach on random data
This is the code which performs a simple regression on some random made up data (the data can be swapped for different data if need be).  The code was inspired by watching YouTube videos online along with taking online Machine Learning courses on sites such as Udacity. 

Feel free to play around with this code yourselves and see if you can find the optimum parameters.